requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
scikit-learn>=1.3.0
xgboost>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
jupyter>=1.0.0
lxml>=4.9.0
psycopg2>=2.9.0
textblob>=0.17.0
python-dotenv 
psycopg2-binary